pyproject.toml
[tool.poetry]
name = "novalug"
version = "0.1.0"
description = ""
authors = ["Bryan Weber <bryan@cobenian.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
xgboost = "^2.1.2"
scikit-learn = "^1.5.2"
black = "^24.10.0"
pandas = "^2.2.3"
streamlit = "^1.40.0"
altair = "^5.4.1"
transformers = "^4.46.2"
datasets = "^3.1.0"
evaluate = "^0.4.3"
accelerate = "^1.1.1"
torch = "^2.5.1"
emoji = "^2.14.0"
termcolor = "^2.5.0"
anthropic = "^0.39.0"
stable-baselines3 = "^2.3.2"
sentence-transformers = "^3.2.1"
llama-index = "^0.11.22"
llama-index-llms-anthropic = "^0.3.9"
numpy = "<2.0.0"
weaviate-client = "^4.9.3"
llama-index-vector-stores-weaviate = "^1.1.3"
llama-index-embeddings-huggingface = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
make-lineup = "novalug.xgb.hitting:main"
learn-to-pitch = "novalug.sb3.learn_to_pitch:main"
practice-calling-pitches = "novalug.sb3.practice_calling_pitches:main"
recommend-similar-hitter = "novalug.sbert.recommend_similar_hitter:main"
call-pitches = "novalug.li.pitching_workflow:main"
summarize-game-recap = "novalug.hf.summarize_game_recap:main"
game-sentiment = "novalug.hf.game_sentiment:main"
chat-about-the-season = "novalug.wvt.rag:main"
agents = "novalug.li.agents:main"